requirements.txt
asgiref==3.7.2
attrs==24.2.0
# backports.zoneinfo==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==0.5.0
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-schema-graph==2.2.1
djangorestframework==3.15.1
idna==3.10
importlib-metadata==6.7.0
mysqlclient==2.1.1
numpy>=2.4.3
pandas==3.0.1
Pillow==9.5.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
six==1.17.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.7
whitenoise==6.5.0
zipp==3.15.0
gunicorn

# Environment & Configuration
python-dotenv==1.2.2

# Data Processing
polars==1.7.1

scipy>=1.11.0
pyarrow>=15.0.0

# AI/ML Integration
google-generativeai==0.8.6
google-api-core==2.15.0
tenacity==9.1.4

# File Format Handling
openpyxl==3.1.5
lxml==6.0.3
pyyaml==6.0.3

python-dateutil>=2.8.2

